+++ core/tasks/bulk_campaign_trigger.go
@@ -9,13 +9,14 @@ import (
 	"github.com/nyaruka/gocommon/dates"
 	"github.com/nyaruka/gocommon/i18n"
 	"github.com/nyaruka/goflow/assets"
+	"github.com/nyaruka/goflow/core/events"
 	"github.com/nyaruka/goflow/flows"
-	"github.com/nyaruka/goflow/flows/events"
 	"github.com/nyaruka/goflow/flows/triggers"
 	"github.com/nyaruka/mailroom/v26/core/ivr"
 	"github.com/nyaruka/mailroom/v26/core/models"
 	"github.com/nyaruka/mailroom/v26/core/runner"
 	"github.com/nyaruka/mailroom/v26/runtime"
+	"github.com/nyaruka/mailroom/v26/utils"
 	"github.com/nyaruka/vkutil"
 )
 
@@ -78,7 +79,7 @@ func (t *BulkCampaignTrigger) Perform(ctx context.Context, rt *runtime.Runtime, 
 
 	for _, cid := range started[:min(recentFiresCap, len(started))] {
 		// set members need to be unique, so we include a random string
-		value := fmt.Sprintf("%s|%d", vkutil.RandomBase64(10), cid)
+		value := fmt.Sprintf("%s|%d", utils.RandomBase64(10), cid)
 		score := float64(dates.Now().UnixNano()) / float64(1e9) // score is UNIX time as floating point
 
 		err := recentSet.Add(ctx, vc, value, score)
@@ -109,7 +110,7 @@ func (t *BulkCampaignTrigger) triggerFlow(ctx context.Context, rt *runtime.Runti
 
 	flowRef := assets.NewFlowReference(flow.UUID(), flow.Name())
 	triggerBuilder := func() flows.Trigger {
-		return triggers.NewBuilder(flowRef).CampaignFired(events.NewCampaignFired(campaign, p.UUID), campaign).Build()
+		return triggers.NewBuilder(flowRef).CampaignFired(events.NewCampaignFired(campaign.Reference(), p.UUID), campaign).Build()
 	}
 
 	if flow.FlowType() == models.FlowTypeVoice {